use of init file
