+++ app/000_init.go
@@ -1,0 +1,3 @@
+package app
+
+import _ "github.com/axelarnetwork/axelar-core/app/encoding_fix"
