--- app/_init.go
@@ -1,3 +1,0 @@
-package app
-
-import _ "github.com/axelarnetwork/axelar-core/app/codec"
